+++ crates/casper-cli/src/main.rs
@@ -1,5 +1,5 @@
-mod arguments;
 mod account;
+mod arguments;
 mod cl_type;
 mod cl_value;
 mod config;